+++ KeireHub/Include/KeireHub/HubChromeLayout.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Keire/Window.h"
+
+namespace KeireHub
+{
+    [[nodiscard]] Keire::WindowChromeLayout BuildHubChromeLayout(Keire::LogicalExtent size,
+                                                                 bool reserveProductControls = true);
+} // namespace KeireHub